# internal/adapters/api/v1/models/models.gen.go
// Package models provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/deepmap/oapi-codegen/v2 version v2.1.1-0.20240427135415-621b44626ac9 DO NOT EDIT.
package models

import (
	"time"
)

// Credential defines model for Credential.
type Credential struct {
	// Login user login
	Login    string `json:"login"`
	Password string `json:"password"`
}

// Data defines model for Data.
type Data struct {
	BankCard    *DataBankCard    `json:"bank_card,omitempty"`
	Binary      *DataBinary      `json:"binary,omitempty"`
	Credentials *DataCredentials `json:"credentials,omitempty"`
	SecretMeta  []SecretMeta     `json:"secret_meta"`
	SecretName  string           `json:"secret_name"`
	SecretType  int32            `json:"secret_type"`
	Text        *DataText        `json:"text,omitempty"`
}

// DataBankCard defines model for DataBankCard.
type DataBankCard struct {
	Holder         string `json:"holder"`
	Number         string `json:"number"`
	Valid          string `json:"valid"`
	ValidationCode string `json:"validation_code"`
}

// DataBinary defines model for DataBinary.
type DataBinary = []byte

// DataCredentials defines model for DataCredentials.
type DataCredentials struct {
	Login    string `json:"login"`
	Password string `json:"password"`
}

// DataText defines model for DataText.
type DataText = string

// Error defines model for Error.
type Error struct {
	Error string `json:"error"`
}

// Secret defines model for Secret.
type Secret struct {
	Created time.Time `json:"created"`
	Id      int32     `json:"id"`
	Name    string    `json:"name"`
	Type    int32     `json:"type"`
	Updated time.Time `json:"updated"`
	UserId  int32     `json:"user_id"`
}

// SecretMeta defines model for SecretMeta.
type SecretMeta struct {
	Attr  string `json:"attr"`
	Value string `json:"value"`
}

// Secrets defines model for Secrets.
type Secrets = []Secret

// SecretListParams defines parameters for SecretList.
type SecretListParams struct {
	// Authorization JWT bearer of user
	Authorization string `form:"Authorization" json:"Authorization"`
}

// SecretSetParams defines parameters for SecretSet.
type SecretSetParams struct {
	// Authorization JWT bearer of user
	Authorization string `form:"Authorization" json:"Authorization"`
}

// SecretGetByIDParams defines parameters for SecretGetByID.
type SecretGetByIDParams struct {
	// Authorization JWT bearer of user
	Authorization string `form:"Authorization" json:"Authorization"`
}

// LoginJSONRequestBody defines body for Login for application/json ContentType.
type LoginJSONRequestBody = Credential

// RegisterJSONRequestBody defines body for Register for application/json ContentType.
type RegisterJSONRequestBody = Credential

// SecretSetJSONRequestBody defines body for SecretSet for application/json ContentType.
type SecretSetJSONRequestBody = Data